feat: reorganize imports and rename components for improved structure and consistency
