+++ src/components/d-flow/data-type/index.ts
@@ -1,0 +1,2 @@
+export * from "./DFlowDataType.service"
+export * from "./DFlowDataType.view"
