+++ src/components/d-flow/edge/DFlowEdge.tsx
@@ -1,7 +1,7 @@
-import {Code0Component} from "../../utils/types";
+import {Code0Component} from "../../../utils/types";
 import {BaseEdge, Edge, EdgeLabelRenderer, EdgeProps, getSmoothStepPath, Position} from "@xyflow/react";
 import React, {memo} from "react";
-import {Badge} from "../badge/Badge";
+import {Badge} from "../../badge/Badge";
 
 export interface DFlowEdgeDataProps extends Code0Component<HTMLDivElement> {
     //some data we will use
