+++ src/components/d-flow/function/index.ts
@@ -1,0 +1,2 @@
+export * from "./DFlowFunction.service"
+export * from "./DFlowFunction.view"
