+++ src/components/d-flow/minimap/DFlowMiniMap.style.scss
@@ -1,6 +1,6 @@
-@use "../../styles/helpers";
-@use "../../styles/box";
-@use "../../styles/variables";
+@use "../../../styles/helpers";
+@use "../../../styles/box";
+@use "../../../styles/variables";
 
 .d-flow-viewport-mini-map {
