+++ src/components/d-flow/minimap/index.ts
@@ -1,0 +1,1 @@
+export * from "./DFlowMiniMap"
