+++ src/components/d-flow/suggestion/index.ts
@@ -1,0 +1,2 @@
+export * from "./DFlowSuggestion.service"
+export * from "./DFlowSuggestion.view"
