+++ src/components/d-flow/type/index.ts
@@ -1,0 +1,2 @@
+export * from "./DFlowType.service"
+export * from "./DFlowType.view"
