+++ src/components/d-flow/validation/DFlowValidation.style.scss
@@ -1,6 +1,6 @@
-@use "../../styles/variables";
-@use "../../styles/helpers";
-@use "../../styles/box";
+@use "../../../styles/variables";
+@use "../../../styles/helpers";
+@use "../../../styles/box";
 
 .d-flow-viewport-validations {
   padding: variables.$xxs;
