+++ src/components/d-organization/index.ts
@@ -1,0 +1,4 @@
+export * from "./DOrganization.service"
+export * from "./DOrganization.view"
+export * from "./DOrganizationCard"
+export * from "./DOrganizationMenu"
