+++ src/components/d-runtime/index.ts
@@ -1,0 +1,2 @@
+export * from "./DRuntime.service"
+export * from "./DRuntime.view"
