+++ src/components/d-user/index.ts
@@ -1,0 +1,3 @@
+export * from "./DUser.service"
+export * from "./DUser.view"
+export * from "./DUserMenu"
