--- src/utils/translation.ts
@@ -1,4 +1,0 @@
-export interface Translation {
-    code: string //de_DE
-    text: string
-}
